+++ CHANGES.md
@@ -9,6 +9,7 @@
 - #1199 Can't get back from to login screen from registration screen
 - #1204 Link encoding issue
 - #1214 Setting `allow_contact_requests` to `false` has no effect
+- #1221 Avoid creating a headlines box if we don't have anything to show inside it
 - #1222 Adding a bookmark should prefill the room name
 
 ## 4.0.2 (2018-10-02)